# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST de finanzas personales",
        "title": "FinApp API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/workspaces/{workspace_id}/debts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "List debts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Debt"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Create debt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Debt data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createDebtRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Debt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/workspaces/{workspace_id}/debts/{debt_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Get debt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "debt_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Debt"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Update debt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "debt_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Debt data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateDebtRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Debt"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Delete debt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "debt_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/workspaces/{workspace_id}/debts/{debt_id}/payments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "List debt payments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "debt_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DebtPayment"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Record debt payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "debt_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.recordPaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.DebtPayment"
                        }
                    }
                }
            }
        },
        "/workspaces/{workspace_id}/debts/{debt_id}/payments/{payment_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Update debt payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "debt_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "payment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DebtPayment"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Delete debt payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "debt_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "payment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/workspaces/{workspace_id}/debts/{debt_id}/schedule": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Get amortization schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "debt_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DebtScheduleInstallment"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.createDebtRequest": {
            "type": "object",
            "required": [
                "first_payment_date",
                "installments",
                "name",
                "principal",
                "rate_type"
            ],
            "properties": {
                "first_payment_date": {
                    "type": "string"
                },
                "installments": {
                    "type": "integer",
                    "minimum": 1
                },
                "lender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "principal": {
                    "type": "number"
                },
                "rate": {
                    "type": "number",
                    "minimum": 0
                },
                "rate_type": {
                    "type": "string"
                }
            }
        },
...
        "handlers.recordPaymentRequest": {
            "type": "object",
            "required": [
                "amount",
                "paid_at",
                "period"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "notes": {
                    "type": "string"
                },
                "paid_at": {
                    "type": "string"
                },
                "period": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "handlers.updateDebtRequest": {
            "type": "object",
            "required": [
                "first_payment_date",
                "installments",
                "name",
                "principal",
                "rate_type"
            ],
            "properties": {
                "first_payment_date": {
                    "type": "string"
                },
                "installments": {
                    "type": "integer",
                    "minimum": 1
                },
                "lender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "principal": {
                    "type": "number"
                },
                "rate": {
                    "type": "number",
                    "minimum": 0
                },
                "rate_type": {
                    "type": "string"
                }
            }
        },
        "handlers.updatePaymentRequest": {
            "type": "object",
            "required": [
                "amount",
                "paid_at"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "notes": {
                    "type": "string"
                },
                "paid_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Debt": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "first_payment_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "installments": {
                    "type": "integer"
                },
                "lender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "principal": {
                    "type": "number"
                },
                "rate": {
                    "type": "number"
                },
                "rate_type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "workspace_id": {
                    "type": "string"
                }
            }
        },
        "models.DebtPayment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "debt_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "paid_at": {
                    "type": "string"
                },
                "period": {
                    "type": "integer"
                }
            }
        },
        "models.DebtScheduleInstallment": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "due_date": {
                    "type": "string"
                },
                "interest": {
                    "type": "number"
                },
                "paid_amount": {
                    "type": "number"
                },
                "paid_at": {
                    "type": "string"
                },
                "payment": {
                    "type": "number"
                },
                "period": {
                    "type": "integer"
                },
                "principal": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}